backend/requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
pymongo==4.6.1
dnspython==2.4.2